updated README for Arranger, added font awesome, moved charts to helper funcs
