+++ src/DataExplorer/DataExplorer.less
@@ -1,3 +1,5 @@
+@import '../css/base.less';
+
 .data-explorer {
   display: flex;
   margin: 30px 30px;
@@ -12,6 +14,16 @@
   flex-flow: column;
   padding-left: 20px;
   width: 80%;
+}
+
+.data-explorer__results-title {
+  display: flex;
+  align-items: center;
+}
+
+.data-explorer__results-title .fa {
+  margin-left: 5px;
+  color: @light_gray;
 }
 
 .data-explorer__visualizations {
