+++ src/DataExplorer/DataExplorerResults.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import FontAwesome from 'react-fontawesome';
 import { CurrentSQON } from '@arranger/components/dist/Arranger';
 import DataExplorerTable from '../components/tables/DataExplorerTable';
 import SummaryChartGroup from '../components/charts/SummaryChartGroup';
 import SummaryHorizontalBarChart from '../components/charts/SummaryHorizontalBarChart';
 import DataSummaryCardGroup from '../components/cards/DataSummaryCardGroup';
+import { getSummaries } from '../components/charts/helper';
 import { localTheme } from '../localconf';
 
 class DataExplorerResults extends React.Component {
@@ -20,64 +22,17 @@ class DataExplorerResults extends React.Component {
     this.setState({ showVisualization: !this.state.showVisualization });
   }
 
-  transformArrangerDataToSummary = (field, chartType, title) => {
-    return {
-      type: chartType,
-      title: title,
-      data: this.transformArrangerDataToChart(field),
-    }
-  }
-
-  transformArrangerDataToChart = (field) => {
-    let chartData = [];
-    field.buckets.map(bucket => {
-      chartData.push({
-        name: bucket.key,
-        value: bucket.doc_count,
-      })
-    })
-    return chartData;
-  }
-
-  transformDataToCount = (field, label) => {
-    return {
-      label: label,
-      value: field.buckets.length,
-    }
-  }
-
-  getSummaries = (data) => {
-    let countItems = [];
-    let charts = [];
-    let horizontalBarCharts = [];
-
-    if (data && data.subject.aggregations) {
-      let fields = data.subject.aggregations;
-      Object.keys(fields).map(field => {
-        let fieldConfig = this.props.arrangerConfig.charts[field]
-        if (fieldConfig) {
-          switch(fieldConfig.chartType) {
-            case 'count':
-              return countItems.push(this.transformDataToCount(fields[field], fieldConfig.title));
-            case 'pie':
-            case 'bar':
-              return charts.push(this.transformArrangerDataToSummary(fields[field], fieldConfig.chartType, fieldConfig.title));
-            case 'horizontalBar':
-              return horizontalBarCharts.push(this.transformArrangerDataToSummary(fields[field], fieldConfig.chartType, fieldConfig.title));
-            default:
-              return;
-          }
-        }
-      })
-    }
-    return { charts: charts, countItems: countItems, horizontalBarCharts: horizontalBarCharts};
-  }
-
   render() {
-    const summaries = this.props.arrangerData ? this.getSummaries(this.props.arrangerData) : null;
+    const summaries = this.props.arrangerData ? getSummaries(this.props.arrangerData, this.props.arrangerConfig) : null;
     return (
       <div className="data-explorer__results">
-        <h4 onClick={this.toggleVisualization}>Data Summary</h4>
+        <div
+          className="data-explorer__results-title"
+          onClick={this.toggleVisualization}
+        >
+          <h4>Data Summary</h4>
+          <FontAwesome name={this.state.showVisualization ? 'chevron-down' : 'chevron-up'} />
+        </div>
         <CurrentSQON className="data-explorer__sqon" {...this.props} />
         { summaries && this.state.showVisualization ?
           <div className="data-explorer__visualizations">
